only loginuser use tasklist

+++ app/Http/Controllers/TasksController.php
@@ -15,9 +15,19 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::all();
+        $data = [];
+        if(\Auth::check()) {//認証済みの場合
+            //認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
         
-        return view('tasks.index', ['tasks' => $tasks, ]);
+        return view('dashboard', $data);
     }
 
     /**
@@ -45,10 +55,10 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status
+        ]);
         
         return redirect('/');
     }
@@ -61,6 +71,7 @@ class TasksController extends Controller
      */
     public function show($id)
     {
+        
         $task = Task::findOrFail($id);
         
         return view('tasks.show', ['task' => $task, ]);
@@ -113,8 +124,15 @@ class TasksController extends Controller
     {
         $task = Task::findOrFail($id);
         
-        $task->delete();
+        if(\Auth::id() === $task->user_id) {
+            $task->delete();
+            return redirect('/')
+                ->with('success', 'Delete Successful');
+        }
         
-        return redirect('/');
+        
+        
+        return redirect('/')
+            ->with('Delete Failed');
     }
 }
